FractalMLServer-Lite/modelhost/flask_app/requirements.txt
click==8.0.1
dataclasses==0.8
Flask==2.0.1
Flask-HTTPAuth==4.4.0
flatbuffers==2.0
gunicorn==20.1.0
importlib-metadata==4.5.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
linux-metrics==0.1.4
MarkupSafe==2.0.1
numpy==1.19.5
onnx==1.9.0
onnxconverter-common==1.8.1
onnxruntime==1.8.0
prometheus-client==0.11.0
protobuf==3.17.3
python-dotenv==0.17.1
scikit-learn==0.24.2
#TODO 0.21.3
scipy==1.5.4
six==1.16.0
skl2onnx==1.8.0
threadpoolctl==2.1.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
zipp==3.4.1